requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Werkzeug==3.1.3
qrcode[pil]==8.0
pycryptodome==3.21.0
